Dejagnu: telnet: do not check command echoed back
When we send some command over telnet, we could expect that we
will see this command echoed back in terminal. And there was a check
that command was echoed back. But if command was longer than 60
symbols, terminal inserted new line symbol and we got modified
command echoed back. So every time we got command longer than 60
symols we had to wait 30 seconds before we get message that echoed
command is not match inserted command. So we just removed this check